+++ packages/pn-pa-webapp/src/pages/Statistics.page.tsx
@@ -11,11 +11,13 @@ import {
   formatDate,
   formatToSlicedISOString,
   getDateFromString,
+  isDateInRange,
   oneYearAgo,
   screenshot,
   today,
 } from '@pagopa-pn/pn-commons';
 import { MIButton } from '@pagopa/mui-italia';
+import { MIAlert } from '@pagopa/mui-italia';
 
 import DeliveryModeStatistics from '../components/Statistics/DeliveryModeStatistics';
 import DigitalErrorsDetailStatistics from '../components/Statistics/DigitalErrorsDetailStatistics';
@@ -32,6 +34,7 @@ import { useAppDispatch, useAppSelector } from '../redux/hooks';
 import { STATISTICS_ACTIONS, getStatistics } from '../redux/statistics/actions';
 import { hasData } from '../redux/statistics/reducers';
 import { RootState } from '../redux/store';
+import { getConfiguration } from '../services/configuration.service';
 import PAEventStrategyFactory from '../utility/MixpanelUtils/PAEventStrategyFactory';
 
 const filter = (node: HTMLElement) => {
@@ -71,6 +74,10 @@ const Statistics = () => {
     (state: RootState) => state.userState.user?.organization
   );
   const isSupportUser = useAppSelector(authSelectors.selectIsSupportUser);
+  const statisticsMaintenanceDates = getConfiguration().STATISTICS_MAINTENANCE_DATES;
+  const [statisticsMaintenanceStartDate, statisticsMaintenanceEndDate] = statisticsMaintenanceDates
+    ? statisticsMaintenanceDates.split('_')
+    : [];
 
   const cxId = loggedUserOrganizationParty.id;
   const cxType = isSupportUser ? CxType.BS : CxType.PA;
@@ -136,6 +143,21 @@ const Statistics = () => {
             <Typography variant="caption" sx={{ color: GraphColors.greyBlue }}>
               {getLastUpdateText()}
             </Typography>
+            {statisticsMaintenanceDates &&
+              isDateInRange(
+                new Date(),
+                statisticsMaintenanceStartDate,
+                statisticsMaintenanceEndDate
+              ) && (
+                <MIAlert
+                  data-testid="maintenanceAlert"
+                  severity="warning"
+                  title={t('maintenance_alert.title')}
+                  sx={{ mt: 4 }}
+                >
+                  {t('maintenance_alert.description')}
+                </MIAlert>
+              )}
             <Box ref={exportJpgNode}>
               <Typography variant="h6" component="h5" mt={7}>
                 {t('section_1')}